pyproject.toml
[tool.poetry]
name = "petstore"
version = "0.0.1"
description = "Pet store API"
authors = ["Mike Bannister <notimpossiblemike@gmail.com>"]
readme = "readme.md"
packages = [{include = "petstore"}]

[tool.poetry.dependencies]
aiohttp = "^3.8.4"
python = "^3.11"
strawberry-graphql = "^0.189.1"
python-ulid = "^1.1.0"

[tool.poetry.group.dev.dependencies]
aiohttp-devtools = "^1.0.post0"
black = "^23.3.0"
flake8 = "^6.0.0"
mypy = "^1.3.0"
pytest = "^7.4.0"
pytest-watch = "^4.2.0"
pytest-asyncio = "^0.21.0"
pytest-aiohttp = "^1.0.4"
gql = {extras = ["aiohttp", "websockets"], version = "^3.4.1"}
coverage = "^7.2.7"
pdoc = "^14.0.0"
rich = "^13.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--verbose --capture=no"

[tool.mypy]
disallow_untyped_defs = true
